Add JsonMockRequestGenerator. It generates mock HTTP requests with JSON body for a given route and an object. (#9)
